pyproject.toml
[project]
name = "factory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.124.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt[crypto]>=2.10.1",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.38.0",
]
